+++ tests/LinkDotNet.StringBuilder.UnitTests/ValueStringBuilderTests.cs
@@ -207,6 +207,16 @@ public class ValueStringBuilderTests
     }
 
     [Fact]
+    public void ShouldReturnZeroIfBothEmpty()
+    {
+        using var stringBuilder = new ValueStringBuilder();
+
+        var index = stringBuilder.IndexOf(string.Empty);
+
+        index.Should().Be(0);
+    }
+
+    [Fact]
     public void ShouldReturnMinusOneWordIsLongerThanString()
     {
         using var stringBuilder = new ValueStringBuilder();
@@ -303,6 +313,16 @@ public class ValueStringBuilderTests
         builder.Append("Hello");
 
         var index = builder.IndexOf(string.Empty, 6);
+
+        index.Should().Be(0);
+    }
+
+    [Fact]
+    public void ShouldReturnZeroIfBothEmptyLastIndexOf()
+    {
+        using var stringBuilder = new ValueStringBuilder();
+
+        var index = stringBuilder.LastIndexOf(string.Empty);
 
         index.Should().Be(0);
     }